requirements.txt
# Brain - 无人系统任务分解大脑
# 依赖清单

# 核心依赖
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
asyncio-compat>=0.1.0

# LLM接口
openai>=1.3.0
anthropic>=0.7.0
httpx>=0.25.0
ollama>=0.3.0  # Ollama本地大模型

# 状态机与任务调度
transitions>=0.9.0
apscheduler>=3.10.0

# 日志与监控
loguru>=0.7.2
prometheus-client>=0.19.0

# 数据处理
numpy>=1.24.0
scipy>=1.11.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0  # 可选，用于图像处理

# VLM/YOLO (可选)
# ultralytics>=8.0.0  # YOLO目标检测，需要时取消注释

# 通信协议
pyzmq>=25.1.0
protobuf>=4.24.0
grpcio>=1.59.0

# 地理空间计算
pyproj>=3.6.0
shapely>=2.0.0

# 可视化调试 (可选)
matplotlib>=3.8.0
networkx>=3.2.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 类型检查
mypy>=1.7.0

# ROS2相关 (需要单独安装ROS2环境后才能使用)
# rclpy  # 通过ROS2安装
# geometry_msgs
# sensor_msgs
# nav_msgs
# cv_bridge

